cleanup pluralization on commands
